+++ src/main/java/ru/sibintek/cis/dao/PmIrDAO.java
@@ -9,9 +9,9 @@ import java.util.List;
 public interface PmIrDAO {
     void delete(PmIrEntity psIr);
 
-    PmIrEntity getById(Long id);
+    PmIrEntity getById(int id);
 
-    void save(PmIrEntity psIr, Long id);
+    void save(PmIrEntity psIr, int id);
 
     List<PmIrEntity> getByCriteria(Criterion criterion);
